requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
async-timeout==4.0.3
bleach==6.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
cmdstanpy==1.2.4
contourpy==1.3.0
crispy-bootstrap4==2024.1
cycler==0.12.1
distro==1.9.0
dj-database-url==2.1.0
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==64.1.0
django-appconf==1.0.6
django-cors-headers==4.3.1
django-crispy-forms==2.3
django-filter==24.3
django-redis==5.4.0
django-select2==8.2.1
django-summernote==0.8.20.0
django-tables2==2.7.0
djangorestframework==3.15.2
exceptiongroup==1.3.0
fonttools==4.54.1
gunicorn==23.0.0
h11==0.16.0
holidays==0.58
httpcore==1.0.9
httpx==0.27.2
idna==3.10
importlib_resources==6.4.5
jiter==0.11.0
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.2
openai==1.55.3
packaging==24.1
pandas==2.2.3
pillow==10.4.0
prophet==1.1.6
psycopg2-binary==2.9.9
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
qrcode==8.0
redis==5.0.8
requests==2.32.3
scikit-fuzzy==0.5.0
scipy==1.14.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stanio==0.5.1
tqdm==4.67.1
typesense==1.1.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1